+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Craft Seeder Changelog
 
-## 3.3.0 - 2029-10-12
+## 3.3.1 - 2029-10-12
 ### Added
-- This update adds some basic options to make the seeding less random and more based on other site content. See the docs [here](https://github.com/studioespresso/craft3-seeder#settings). 
+- Added support for Craft's number field
+- Plain text fields now work with the fields config
+- Titles based in custom fields nowget set correctly 
+
 
 ## 3.2.0 - 2019-12-23
 ### Added